Move OptionExplicitInspectionTests for QuickFixes into relevant files under RD.QuickFixes
