+++ RubberduckTests/QuickFixes/OptionExplicitQuickFixTests.cs
@@ -1,0 +1,38 @@
+﻿using System.Linq;
+using System.Threading;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+using Rubberduck.Inspections.Concrete;
+using Rubberduck.Inspections.QuickFixes;
+using Rubberduck.VBEditor.SafeComWrappers.Abstract;
+using RubberduckTests.Mocks;
+using RubberduckTests.Inspections;
+
+namespace RubberduckTests.QuickFixes
+{
+    [TestClass]
+    public class OptionExplicitQuickFixTests
+    {
+        [TestMethod]
+        [TestCategory("QuickFixes")]
+        public void NotAlreadySpecified_QuickFixWorks()
+        {
+            const string inputCode = "";
+            const string expectedCode =
+@"Option Explicit
+
+";
+
+            IVBComponent component;
+            var vbe = MockVbeBuilder.BuildFromSingleStandardModule(inputCode, out component);
+            var state = MockParser.CreateAndParse(vbe.Object);
+
+            var inspection = new OptionExplicitInspection(state);
+            var inspector = InspectionsHelper.GetInspector(inspection);
+            var inspectionResults = inspector.FindIssuesAsync(state, CancellationToken.None).Result;
+
+            new OptionExplicitQuickFix(state).Fix(inspectionResults.First());
+            Assert.AreEqual(expectedCode, state.GetRewriter(component).GetText());
+        }
+
+    }
+}
