+++ RubberduckTests/RubberduckTests.csproj
@@ -125,6 +125,7 @@
     <Compile Include="QuickFixes\IgnoreOnceQuickFixTests.cs" />
     <Compile Include="QuickFixes\MakeSingleLineParamegerQuickFixTests.cs" />
     <Compile Include="QuickFixes\MoveFieldCloserToUsageQuickFixTests.cs" />
+    <Compile Include="QuickFixes\OptionExplicitQuickFixTests.cs" />
     <Compile Include="QuickFixes\RemoveCommentQuickFixTests.cs" />
     <Compile Include="QuickFixes\RemoveEmptyElseBlockQuickFixTests.cs" />
     <Compile Include="QuickFixes\RemoveEmptyIfBlockQuickFixTests.cs" />
